requirements.txt
aiofiles==24.1.0
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
piexif==1.1.3
pillow==11.1.0
typing_extensions==4.12.2
Werkzeug==3.1.3
